Improve heartbeat check, force to close connection when connection timeout (#3563)
* force to close connection

* ignore conn->closed

+++ src/protocol/base.cc
@@ -86,7 +86,7 @@ int Protocol::recv_split_by_eof(network::Socket *socket, String *buffer) {
         return retval;
     } else if (n == 0) {
         return SW_CONTINUE;
-    } else if (n < buffer->length) {
+    } else if (n < (ssize_t) buffer->length) {
         off_t offset;
         buffer->reduce(n);
         offset = buffer->length - package_eof_len;
